requirements.txt
# Required dependencies for sshPilot
# System requirements (install via package manager):
# - python>=3.10
# - pygobject>=3.42 (python3-gi)
# - libadwaita>=1.2 (gir1.2-adw-1)
# - vte>=0.70 (gir1.2-vte-2.91-gtk4)

# Python packages (installable via pip):
paramiko>=3.4
pyyaml>=6.0
secretstorage>=3.3
cryptography>=42.0
matplotlib>=3.8

# Development dependencies (optional):
# pytest>=7.0
# pytest-cov>=4.0
# black>=23.0
# flake8>=6.0
# mypy>=1.0